requirements.txt
# Core ML stack — tested on RTX 4090 (Ada, sm_89) Runpod.
#
# IMPORTANT: do NOT pip-install `torch` from this requirements file directly.
# Use scripts/run_runpod.sh which pins torch==2.6.0 from the cu124 index:
#
#   pip install --force-reinstall --index-url https://download.pytorch.org/whl/cu124 torch==2.6.0
#
# Why this exact version:
#   - PyPI default index serves a CPU-only torch which won't see your GPU.
#   - cu128 index serves +cu130 wheels at HEAD (need driver 12.8+, Runpod
#     4090 templates ship driver 12.4).
#   - transformers >= 4.51 added check_torch_load_is_safe() (CVE-2025-32434);
#     it refuses .bin loads on torch < 2.6 — and microsoft/longcoder-base is
#     .bin only (no safetensors mirror on HF). 2.6 is therefore the floor.
#   - torch 2.10+ caps setuptools<82, which breaks modern pip envs.
#   - torch 2.7+ bundles NCCL 2.21+ and references ncclCommWindowDeregister,
#     which can hit "undefined symbol" errors if a base image's older system
#     libnccl shadow-loads ahead of torch's bundled one. 2.6.0 bundles NCCL
#     2.20.x, sidestepping that whole class of issue.
#
# If the base image ships torchvision/torchaudio, uninstall them to avoid
# transformers' lazy-vision import exploding:
#   pip uninstall -y torchvision torchaudio
#
# We don't list torch here because pip-resolving against the default PyPI
# index would silently grab the CPU build. The bootstrap script does it.
# torch==2.6.0   # installed by scripts/run_runpod.sh, not from this file
transformers>=4.45.0,<5.0.0
peft>=0.13.0
accelerate>=1.0.0
datasets>=2.20.0
safetensors>=0.4.5

# Tokenization / parsing — per-language tree-sitter packages (Python 3.13 abi3 wheels).
# We previously used tree_sitter_languages but it doesn't ship 3.13 wheels.
tree_sitter>=0.23,<1.0
tree-sitter-python>=0.21
tree-sitter-java>=0.21
tree-sitter-cpp>=0.22
tree-sitter-c>=0.22
tree-sitter-c-sharp>=0.21
tree-sitter-go>=0.22
tree-sitter-javascript>=0.22
tree-sitter-typescript>=0.21
tree-sitter-php>=0.22
tree-sitter-ruby>=0.21
tree-sitter-rust>=0.21
tree-sitter-swift>=0.0.1
markdown-it-py>=3.0.0

# Data
pandas>=2.2.0
pyarrow>=16.0.0
numpy>=1.26.0
datasketch>=1.6.5
scikit-learn>=1.4.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
click>=8.1.0
joblib>=1.4.0

# Logging (optional; code falls back to JSONL/TensorBoard if wandb missing)
wandb>=0.17.0; python_version >= "3.9"
tensorboard>=2.15.0

# Plotting + HP search
matplotlib>=3.8.0
optuna>=3.6.0

# Stacking heads (optional)
lightgbm>=4.3.0
xgboost>=2.0.0
